relocate extractSingleFileFromTar function to right file
