+++ test/pkg/compose/helpers/composer_dryrun.go
@@ -1,10 +1,6 @@
 package helpers
 
 import (
-	"archive/tar"
-	"fmt"
-	"io"
-	"os"
 	"strings"
 	"time"
 
@@ -76,47 +72,5 @@ func (c *dryrunComposer) WaitForCompose(id, friendlyName string, timeout time.Du
 	klog.InfoS("DRYRUN: Waiting for compose", "id", id, "timeout", timeout)
 	time.Sleep(1 * time.Second)
 	klog.InfoS("DRYRUN: Waited for compose", "id", id, "timeout", timeout)
-	return nil
-}
-
-// extractSingleFileFromTar extracts exactly one file from tar archive.
-// Intended to be used with compose metadata or logs as the archives contain only single file.
-func extractSingleFileFromTar(archivePath, filePath string) error {
-	archiveFile, err := os.Open(archivePath)
-	if err != nil {
-		return fmt.Errorf("failed to open tar file %q: %w", archivePath, err)
-	}
-	tarFile := tar.NewReader(archiveFile)
-	header, err := tarFile.Next()
-	if err == io.EOF {
-		return fmt.Errorf("tar file %q is empty, expected single file", archivePath)
-	}
-	if err != nil {
-		return fmt.Errorf("error when reading tar file %q: %w", archivePath, err)
-	}
-	if header.Typeflag != tar.TypeReg {
-		return fmt.Errorf("unexpected header type in tar file %q - type: %v", archivePath, header.Typeflag)
-	}
-
-	extractedFile, err := os.OpenFile(filePath, os.O_CREATE|os.O_RDWR, os.FileMode(header.Mode))
-	if err != nil {
-		return fmt.Errorf("failed to open file %q: %w", filePath, err)
-	}
-	defer extractedFile.Close()
-
-	written, err := io.Copy(extractedFile, tarFile)
-	if err != nil {
-		return fmt.Errorf("failed to copy file to %q: %w", filePath, err)
-	}
-
-	if header.Size != written {
-		return fmt.Errorf("incomplete file copy from the tar archive, size:%d, copied:%d", header.Size, written)
-	}
-
-	err = os.Remove(archivePath)
-	if err != nil {
-		return fmt.Errorf("failed to remove old tar file %q: %w", archivePath, err)
-	}
-
 	return nil
 }
